+++ packages/appkit/src/client/appkit.ts
@@ -147,7 +147,6 @@ export class AppKit extends AppKitBaseClient {
           chainNamespace: namespace
         })
       }
-      this.setCaipAddress(caipAddress, namespace)
 
       const { address: smartAccountAddress } =
         user.accounts?.find(account => account.type === 'smartAccount') ?? {}
@@ -155,6 +154,8 @@ export class AppKit extends AppKitBaseClient {
       if (smartAccountAddress && namespace === ConstantsUtil.CHAIN.EVM) {
         this.setSmartAccountAddress(smartAccountAddress, ConstantsUtil.CHAIN.EVM)
       }
+
+      this.setCaipAddress(caipAddress, namespace)
       this.setUser({ ...(AccountController.state.user || {}), ...user }, namespace)
       this.setSmartAccountDeployed(Boolean(user.smartAccountDeployed), namespace)
       this.setPreferredAccountType(preferredAccountType, namespace)
@@ -203,10 +204,7 @@ export class AppKit extends AppKitBaseClient {
     const email = provider.getEmail()
     const username = provider.getUsername()
 
-    this.setUser(
-      { ...(AccountController.state?.user || {}), username, email },
-      ChainController.state.activeChain
-    )
+    this.setUser({ ...(AccountController.state?.user || {}), username, email }, chainNamespace)
 
     this.setupAuthConnectorListeners(provider)
 
@@ -466,7 +464,6 @@ export class AppKit extends AppKitBaseClient {
     await super.initialize(options)
     this.chainNamespaces?.forEach(namespace => {
       this.createAuthProviderForAdapter(namespace)
-      this.chainAdapters?.[namespace].syncConnectors(this.options, this)
     })
     await this.injectModalUi()
     PublicStateController.set({ initialized: true })