+++ packages/controllers/src/controllers/ModalController.ts
@@ -1,9 +1,10 @@
 import { proxy, subscribe as sub } from 'valtio/vanilla'
 import { subscribeKey as subKey } from 'valtio/vanilla/utils'
 
-import type { ChainNamespace } from '@reown/appkit-common'
+import { type ChainNamespace } from '@reown/appkit-common'
 
 import { CoreHelperUtil } from '../utils/CoreHelperUtil.js'
+import { NetworkUtil } from '../utils/NetworkUtil.js'
 import { withErrorBoundary } from '../utils/withErrorBoundary.js'
 import { ApiController } from './ApiController.js'
 import { ChainController } from './ChainController.js'
@@ -56,6 +57,11 @@ const controller = {
   },
 
   async open(options?: ModalControllerArguments['open']) {
+    const namespace = options?.namespace
+    const currentNamespace = ChainController.state.activeChain
+    const isSwitchingNamespace = namespace && namespace !== currentNamespace
+    const caipAddress = ChainController.getAccountData(options?.namespace)?.caipAddress
+
     if (ConnectionController.state.wcBasic) {
       // No need to add an await here if we are use basic
       ApiController.prefetch({ fetchNetworkImages: false, fetchConnectorImages: false })
@@ -63,29 +69,33 @@ const controller = {
       await ApiController.prefetch()
     }
 
-    if (options?.namespace) {
-      await ChainController.switchActiveNamespace(options.namespace)
-      ModalController.setLoading(true, options.namespace)
-    } else {
-      ModalController.setLoading(true)
-    }
     ConnectorController.setFilterByNamespace(options?.namespace)
+    ModalController.setLoading(true, namespace)
 
-    const caipAddress = ChainController.getAccountData(options?.namespace)?.caipAddress
-    const hasNoAdapters = ChainController.state.noAdapters
+    if (namespace && isSwitchingNamespace) {
+      const namespaceNetwork =
+        ChainController.getNetworkData(namespace)?.caipNetwork ||
+        ChainController.getRequestedCaipNetworks(namespace as ChainNamespace)[0]
 
-    if (OptionsController.state.manualWCControl || (hasNoAdapters && !caipAddress)) {
-      if (CoreHelperUtil.isMobile()) {
-        RouterController.reset('AllWallets')
-      } else {
-        RouterController.reset('ConnectingWalletConnectBasic')
+      if (namespaceNetwork) {
+        NetworkUtil.onSwitchNetwork({ network: namespaceNetwork, ignoreSwitchConfirmation: true })
       }
-    } else if (options?.view) {
-      RouterController.reset(options.view, options.data)
-    } else if (caipAddress) {
-      RouterController.reset('Account')
     } else {
-      RouterController.reset('Connect')
+      const hasNoAdapters = ChainController.state.noAdapters
+
+      if (OptionsController.state.manualWCControl || (hasNoAdapters && !caipAddress)) {
+        if (CoreHelperUtil.isMobile()) {
+          RouterController.reset('AllWallets')
+        } else {
+          RouterController.reset('ConnectingWalletConnectBasic')
+        }
+      } else if (options?.view) {
+        RouterController.reset(options.view, options.data)
+      } else if (caipAddress) {
+        RouterController.reset('Account')
+      } else {
+        RouterController.reset('Connect')
+      }
     }
 
     state.open = true
@@ -111,6 +121,7 @@ const controller = {
     }
 
     state.open = false
+    RouterController.reset('Connect')
     ModalController.clearLoading()
 
     if (isEmbeddedEnabled) {