+++ packages/controllers/src/utils/SIWXUtil.ts
@@ -57,7 +57,7 @@ export const SIWXUtil = {
       })
 
       // eslint-disable-next-line no-console
-      await ConnectionController._getClient()?.disconnect().catch(console.error)
+      await ConnectionController._getClient()?.disconnect({}).catch(console.error)
       RouterController.reset('Connect')
       SnackController.showError('A problem occurred while trying initialize authentication')
     }