+++ packages/scaffold-ui/src/views/w3m-networks-view/index.ts
@@ -11,6 +11,7 @@ import {
   ConnectorController,
   CoreHelperUtil,
   EventsController,
+  NetworkUtil,
   RouterController
 } from '@reown/appkit-controllers'
 import { customElement } from '@reown/appkit-ui'
@@ -158,50 +159,7 @@ export class W3mNetworksView extends LitElement {
   }
 
   private onSwitchNetwork(network: CaipNetwork) {
-    const routerData = RouterController.state.data
-    const isSameNetwork = network.id === this.network?.id
-
-    if (isSameNetwork) {
-      return
-    }
-
-    const isDifferentNamespace = network.chainNamespace !== ChainController.state.activeChain
-    const isCurrentNamespaceConnected = AccountController.state.caipAddress
-    const isNextNamespaceConnected = AccountController.getCaipAddress(network.chainNamespace)
-    const connectorId = ConnectorController.getConnectorId(ChainController.state.activeChain)
-
-    /**
-     * If the network is supported by the auth connector, we don't need to show switch active chain view.
-     * But there are some cases like switching from Ethereum to Bitcoin where Bitcoin is not supported by the auth connector and users should connect with another connector.
-     */
-    const isConnectedWithAuth = connectorId === ConstantsUtil.CONNECTOR_ID.AUTH
-    const isSupportedForAuthConnector = ConstantsUtil.AUTH_CONNECTOR_SUPPORTED_CHAINS.find(
-      c => c === network.chainNamespace
-    )
-
-    if (isCurrentNamespaceConnected) {
-      if (isConnectedWithAuth && isSupportedForAuthConnector) {
-        // If user connected with auth connector and the next network is supported by the auth connector, we don't need to show switch active chain view.
-        RouterController.push('SwitchNetwork', { ...routerData, network })
-      } else if (
-        // 1. If user connected with auth connector and the next network is not supported by the auth connector, we need to show switch active chain view.
-        (isConnectedWithAuth && !isSupportedForAuthConnector) ||
-        // 2. If user connected with non-auth connector, we should check if user switching to a different namespace and next namespace is not connected.
-        (isDifferentNamespace && !isNextNamespaceConnected)
-      ) {
-        RouterController.push('SwitchActiveChain', {
-          switchToChain: network.chainNamespace,
-          navigateTo: 'Connect',
-          navigateWithReplace: true,
-          network
-        })
-      } else {
-        // For any other case, we redirect to switch network page
-        RouterController.push('SwitchNetwork', { ...routerData, network })
-      }
-    } else {
-      RouterController.push('SwitchNetwork', { ...routerData, network })
-    }
+    NetworkUtil.onSwitchNetwork({ network })
   }
 }